Revert "Added function for filter by tags and formatted my files with prettier"
